Load Berlin points onto map from API
