+++ App.js
@@ -1,12 +1,14 @@
 import React, { Component } from 'react';
-import MapItem from './MapItem.js';
+import MapContainer from './MapContainer.js';
+
+const url = 'https://api.openchargemap.io/v3/poi/?output=json&countrycode=DE&maxresults=10&compact=true&verbose=false&latitude=52.520008&longitude=13.404954&distance=10&distanceunit=KM';
 
 class App extends Component{
    render(){
       return(
          <div>
             <h1>Hello World</h1>
-            <MapItem />
+            <MapContainer mapData={url}/>
          </div>
       );
    }
